requirements.txt
streamlit>=1.36
python-dotenv>=1.0.1
pydantic>=2.7
openai-agents>=0.0.12
